refactor: openshift-maven-plugin inherits from kubernetes-maven-plugin
Removed code duplicity in openshift-maven-plugin
