+++ openshift-maven-plugin/plugin/src/main/java/org/eclipse/jkube/maven/plugin/mojo/Openshift.java
@@ -1,0 +1,22 @@
+/**
+ * Copyright (c) 2019 Red Hat, Inc.
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at:
+ *
+ *     https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *   Red Hat, Inc. - initial API and implementation
+ */
+package org.eclipse.jkube.maven.plugin.mojo;
+
+public class Openshift {
+
+  public static final String DEFAULT_LOG_PREFIX = "oc:";
+
+  private Openshift() {}
+
+}
