+++ openshift-maven-plugin/plugin/src/main/java/org/eclipse/jkube/maven/plugin/mojo/develop/OpenshiftDeployMojo.java
@@ -13,13 +13,12 @@
  */
 package org.eclipse.jkube.maven.plugin.mojo.develop;
 
-import org.eclipse.jkube.maven.plugin.mojo.build.ApplyMojo;
-import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.Execute;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
-import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.plugins.annotations.ResolutionScope;
+
+import static org.eclipse.jkube.maven.plugin.mojo.Openshift.DEFAULT_LOG_PREFIX;
 
 /**
  * This goal forks the install goal then applies the generated kubernetes resources to the current cluster.
@@ -32,16 +31,10 @@ import org.apache.maven.plugins.annotations.ResolutionScope;
 
 @Mojo(name = "deploy", requiresDependencyResolution = ResolutionScope.COMPILE, defaultPhase = LifecyclePhase.VALIDATE)
 @Execute(phase = LifecyclePhase.INSTALL)
-public class DeployMojo extends ApplyMojo {
-    @Parameter(property = "docker.skip.deploy", defaultValue = "false")
-    protected boolean skipDeploy;
+public class OpenshiftDeployMojo extends DeployMojo {
 
     @Override
-    public void executeInternal() throws MojoExecutionException {
-        if (skipDeploy) {
-            return;
-        }
-
-        super.executeInternal();
+    protected String getLogPrefix() {
+        return DEFAULT_LOG_PREFIX;
     }
 }
