+++ quickstarts/maven/ibm-javaee8-webprofile-liberty/javaee8-webprofile-liberty-app-it/pom.xml
@@ -21,12 +21,12 @@
     <parent>
         <groupId>org.eclipse.jkube.quickstarts.maven</groupId>
         <artifactId>javaee8-webprofile-liberty</artifactId>
-        <version>0.2-SNAPSHOT</version>
+        <version>0.2.1-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
     <artifactId>javaee8-webprofile-liberty-app-it</artifactId>
-    <version>0.2-SNAPSHOT</version>
+    <version>0.2.1-SNAPSHOT</version>
     <name>Eclipse JKube :: Quickstarts :: Maven :: JavaEE 8 WebProfile :: IT</name>
     <packaging>jar</packaging>
