+++ quickstarts/maven/ibm-javaee8-webprofile-liberty/pom.xml
@@ -20,7 +20,7 @@
 
     <groupId>org.eclipse.jkube.quickstarts.maven</groupId>
     <artifactId>javaee8-webprofile-liberty</artifactId>
-    <version>0.2-SNAPSHOT</version>
+    <version>0.2.1-SNAPSHOT</version>
     <name>Eclipse JKube :: Quickstarts :: Maven :: JavaEE 8 WebProfile</name>
     <packaging>pom</packaging>
