+++ quickstarts/maven/openliberty/pom.xml
@@ -20,7 +20,7 @@
 
     <groupId>org.eclipse.jkube.quickstarts.maven</groupId>
     <artifactId>openliberty</artifactId>
-    <version>0.2-SNAPSHOT</version>
+    <version>0.2.1-SNAPSHOT</version>
     <name>Eclipse JKube :: Quickstarts :: Maven :: Open Liberty</name>
     <packaging>war</packaging>
