+++ quickstarts/maven/openliberty/src/main/liberty/config/server.xml
@@ -13,15 +13,15 @@
       Red Hat, Inc. - initial API and implementation
 
 -->
-<server description="Open Liberty Fabric8 Example Server">
+<server description="Open Liberty JKube Example Server">
 
   <featureManager>
-      <feature>jaxrs-2.1</feature>    
+      <feature>jaxrs-2.1</feature>
   </featureManager>
 
   <httpEndpoint httpPort="${default.http.port}" httpsPort="${default.https.port}"
                 id="defaultHttpEndpoint" host="*" />
 
   <webApplication location="openliberty-sample.war" contextRoot="${app.context.root}"/>
-   
+
 </server>
