+++ quickstarts/maven/zero-config/pom.xml
@@ -20,7 +20,7 @@
 
   <artifactId>jkube-maven-sample-zero-config</artifactId>
   <groupId>org.eclipse.jkube</groupId>
-  <version>0.2-SNAPSHOT</version>
+  <version>0.2.1-SNAPSHOT</version>
   <packaging>jar</packaging>
 
   <parent>
